+++ ShadowPack/OptPageBPBMedia.h
@@ -26,8 +26,8 @@ public:
 	CComboBox m_ctlCrypto;
 	INT m_nCrypto;
 	INT m_nBytePerBlock;
-	UINT m_nTotalBlocks;
-	UINT m_nUsedBytes;
+	ULONGLONG m_nTotalBlocks;
+	ULONGLONG m_nUsedBytes;
 	UINT m_nHeaderSize;
 	CString m_strBlockUnit;
 protected:
